added comments and more test to DocService.java

+++ src/main/java/docSharing/CRDT/CRDT.java
@@ -8,7 +8,7 @@ import java.util.Map;
 //This class represents the crdt doc tree data structure.
 public class CRDT {
 
-    public static final double BASE = 5;       //tree root has 2^5 children
+    public static final int BASE = 5;       //tree root has 2^5 children
     public static final int DOC_BEGIN = 0;
 
     public static final int DOC_END = (int)Math.pow(2,BASE) - 1;
